pyproject.toml
[project]
name = "modulo-agendamento-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "fastapi[standard]>=0.115.12",
    "isort>=6.0.1",
    "langchain-core>=0.3.65",
    "langchain-openai>=0.3.23",
    "langgraph>=0.4.8",
    "langsmith>=0.3.45",
    "langgraph-cli[inmem]>=0.2.10",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "psycopg-pool>=3.2.6",
    "psycopg[binary]>=3.2.9",
    "langgraph-checkpoint-postgres>=2.0.21",
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.41",
    "psycopg2-binary>=2.9.10",
]

[tool.black]
line-length = 89
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 89
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
